pyproject.toml
[project]
name = "tuuziane"
description = "Marketplace for peoples"
readme = "README.md"
authors = [
  { name = "stockman", email = "allan.rugano@ubuviz.com" },
]

requires-python = ">=3.13"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.13",
]
version = "0.1.0"
dependencies = [
  "dj-database-url",
  "django",
  "django-environ",
  "django-health-check",
  "django-oscar[sorl-thumbnail]",
  "gunicorn",
  "psutil",
  "psycopg2-binary",
  "pycountry",
  "pysolr",
  "wagtail",
  "django-vite",
  "django-webpack-loader"
]

urls.Documentation = "https://unicef.github.io/hope-country-report"
urls.Issues = "https://github.com/srugano/tuuziane/issues"
urls.Source = "https://github.com/srugano/tuuziane"

[project.optional-dependencies]
dev = [
  "pre-commit",
  "pytest-xdist",
  "pytest",
  "pytest-django",
  "ruff",
  "hatch",
  "hatch-fancy-pypi-readme",
  "mypy",
  "django-stubs",
  "django-stubs-ext",
  "djangorestframework-stubs",
  "faker",
  "freezegun",
  "pytest-cov",
  "pytest-rerunfailures",
  "pytest-randomly",
  "tox",
  "tox-uv",
  "types-freezegun",
  "types-python-dateutil",
  "types-pytz",
  "types-redis",
  "types-requests",
  "types-setuptools",
]

docs = [
    "mkdocs",
]

mypy = [
  "djangorestframework-stubs",
  "lxml-stubs",
  "mypy",
  "types-defusedxml",
  "types-gevent",
  "types-html5lib",
  "types-jmespath",
  "types-markdown",
  "types-psycopg2",
  "types-pygments",
  "types-python-dateutil",
  "types-pytz",
  "types-requests",
]

[tool.ruff]
target-version = "py313"
line-length = 120
indent-width = 4

exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
  "migrations"
]

format.indent-style = "space"
format.quote-style = "double"
format.line-ending = "auto"
format.skip-magic-trailing-comma = false
format.docstring-code-line-length = "dynamic"
format.docstring-code-format = false
lint.select = ["E", "W", "F"]
lint.ignore = [ "F403" ]
lint.fixable = [ "ALL" ]
lint.unfixable = [  ]
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"



[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tuuziane.settings.dev"

[build-system]
requires = ["hatchling", "hatch-fancy-pypi-readme"]
build-backend = "hatchling.build"

[tool.hatch.metadata.hooks.fancy-pypi-readme]
content-type = "text/markdown"
fragments = [
    { path = "README.md" }
]

[tool.uv]
package = true
